clippy

+++ src/_003.rs
@@ -4,7 +4,7 @@ pub struct Solution;
 
 impl Solution {
     pub fn length_of_longest_substring(s: String) -> i32 {
-        let mut last_occur_map = vec![-1 as i32; 256];
+        let mut last_occur_map = vec![-1_i32; 256];
         let mut ans = 0;
         let mut cur = 0;
         for (i, c) in s.chars().enumerate() {
